# 8256318: AArch64: Add support for floating-point absolute difference