--- a/test/hotspot/jtreg/runtime/invokedynamic/TestDynamicConstant.java
+++ b/test/hotspot/jtreg/runtime/invokedynamic/TestDynamicConstant.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Huawei Technologies Co. Ltd. All rights reserved.
+ * Copyright (c) 2020, Huawei Technologies Co., Ltd. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it