@dgbo 2020-11-16
> Looks good. Please add any new instructions to aarch64-asmtest.py and regenerate assembler_aarch64.cpp.

Done, added tests for `fabd` scalar/vector instructions in this script and regenerated the code. 
Verfied with linux-aarch64-server-fastdebug build.