@dgbo 2020-11-17
> _Mailing list message from [Andrew Haley](mailto:aph@redhat.com) on [hotspot-dev](mailto:hotspot-dev@openjdk.java.net):_
> 
> On 16/11/2020 13:31, dongbo (E) wrote:
> 
> > On 2020/11/16 19:57, Andrew Haley wrote:
> > > On 11/16/20 3:15 AM, Ningsheng Jian wrote:
> > > > On Mon, 16 Nov 2020 02:51:05 GMT, Dong Bo <dongbo at openjdk.org> wrote:
> > > > > But I feel a little bit inconsistent that only the new fabd is added into aarch64_neon.ad,
> > > > > while other NEON intructions (i.e. fabs, fsub, fdiv, fsqrt, etc) are still in aarch64.ad.
> > > > > And moving them all from aarch64.ad to aarch64_neon.ad deviates far away from this patch.
> > > > > Yes, I think when we introduced aarch64_neon.ad (m4), we just tried to keep that patch simple and would move other vector rules in future patches. Maybe Andrew can comment on this?
> > > > > I wonder: I'm not sure if we should do the lot in one bang.
> > > 
> > > 
> > > I can't quite figure out the best thing to do. It'll be tricky to move
> > > all of the SIMD instructions, but I guess it's the best thing to do. It'll
> > > make backports hard, but we don't see many in this area.
> > 
> > 
> > Yes. Moving all NEON instructions to aarch64_neon.ad would make the code
> > clearer and more consistent.
> 
> OK. This should definitely be done as a stand-alone change/
> 

We can do it with a new PR. 

@theRealAph @nsjian Is there any further suggestions for this FABD patch?